htaccess and Gzipping
Implemented the .htaccess file for apache servers to enable gzip
compression to boost page speed. This is set to force gzip compression
for linux servers that do not have it enabled. Updated the gulp file to
copy the .htaccess file into the dist directory.
Updated script tags in the html files to call the async method to stop
render blocking scripts on page load.