More Gulp Features
Refactored how html-replace is used, set more relevant titles in the
build headers.
Restructured the paths in the gulp file to be more concise and easier to
implement when creating tasks.
Added a script for google web fonts, allowing .woff files to be created
aiding in reducing render blocking css, allowing the fonts to be cached
from the site directory and allowing them to be served from a site
controlled location.